pyproject.toml
[project]
name = "edinet-client"
version = "0.1.0"
description = "EDINET APIを使用した財務情報取得システム"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "lxml>=4.9.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
